Add files via upload
Clients began to leave Beta Bank. Bank marketers have calculated that it is cheaper to retain current customers than to attract new ones.
It is necessary to predict whether the client will leave the bank in the near future or not. You are provided with historical data on customer behavior and termination of contracts with the bank.
Build a model with an extremely large F1-measure value. To pass the project successfully, you need to bring the metric to 0.59. Check the F1-measure on the test sample yourself.
Additionally, measure AUC-ROC and compare its value with the F1 measure.