Add files via upload
At your disposal is data on the behavior of customers who have already switched to these data plans- "smart" and "ultra". We need to build a model for the classification task that will select the appropriate tariff.

Build a model with the highest possible accuracy value. To pass the project successfully, you need to bring the percentage of correct answers to at least 0.75. Check accuracy on a test sample yourself.